Add the concept of versions to rpc-light
Add the line:

let version="<version>"

to your interface definition file, and this will autogenerate
a 'get_version' RPC call to the client and server. This will
_not_ be present in the server functor signature. The idea
is that the client calls this immediately and asserts that
the result is identical to its own binding of 'version'. If
this fails, the client can then try to find the correct server
version on a different URI.

Signed-off-by: Jon Ludlam <jonathan.ludlam@eu.citrix.com>